refactor: 세션 관리 클래스를 repository에서 manager로 변경
